Work in progress on reducing Smalltalk AST memory footprint
At the moment the parser retains all literal value tokens. This isn't
really necessary once the literal value nodes have been built, and for
methods that contain very large literal byte arrays (for example) this is
costly. We also need the parsing to be as efficient as possible since it
will soon become the front end for the actual compiler.