Switch Transcript from RichTextEdit to MultilineTextEdit
The RTE will not display supplementary Unicode chars correctly unless the
character format is explicitly set, which is awkward. The original reason
to use the RichEdit control for the Transcript was to remove the text limit
on Win95/98 (the normal multiline text edit control was limited to 32Kb),
but clearly this no longer applies so there is no reason to continue to use
it.