Add initial support for binding reference syntax
Binding references are a way to provide late binding of fully qualified
names, and are a prerequisite for adding full Namespace support.

- Add styling and auto-completion for literal binding references
- Also fix styling/auto-completion of brace array.
- Literal BindingReferences have a class context in which they initiate
their search. This is nil for workspace and file-in expressions. Needs to
be included in equality comparison.
- BindingReferences cache the resolved binding (needs further work to
clear this when needed)
- Add meta BindingReferences as this is useful for referencing the class
side, e.g. for method names.
- BindingReferences as names in Packages rather than Symbols
- Initial Refactoring Browser updates for BindingReference use. The RB will
need to be made namespace aware. As a start, replace some obvious
dependencies on the global Smalltalk dictionary and unqualified names.
- Replace direct lookup in Smalltalk from SUnit with use of BindingReference
- Use binding references instead of direct lookups in Smalltalk dictionary
- Fix tests, mostly broken by compiling string literals as Utf8 by default

May need another boot image update, in which case I will squash this.