Push up AnsiString>>includes: to String
Having added support for UTF Strings to primitiveStringNextIndexOfFromTo
in cc3d008, we can now define efficient implementations of `includes:` and
`identityIncludes:` in terms of `nextIndexOf:from:to:` and
`nextIdentityIndexOf:from:to:` (respectively) that will work for UTF
encoded Strings too.

There was already an adequate test of String>>includes: covering the cases
of searching for a multi-codeunit codepoint in UTF Strings, and I've
added a String>>identityIncludes: test based on this.