Improve behaviour of search presenter with namespaces
The search presenter wasn't working very well for namespaced classes
because there can now be a match against the unqualified name, or the
qualified name. It probably needs more work, but this is a significant
improvement and will now display the right matches for qualified class
names, and then all the methods for that class when appropriate.