Add support for implicit custom namespaces for methods
Support compilation of methods in an alternate namespace. This allows
the extension methods of legacy packages to continue to be compiled in
the Smalltalk context so that they can be loaded without requiring
Smalltalk qualified names if the methods are now extending classes that
are not in the Smalltalk namespace.
The mechanism also provides the infrastructure to support explicit method
namespaces, although the compiler support for parsing namespace tags is
not implemented yet.
Quite a lot of changes are required to achieve this, and the commit also
includes some changes that revert the addition of Smalltalk-qualified names
in the extension methods of legacy packages.