Fix #258: Metaclass identity not preserved by ClassBuilder modifying class
This bug was reported in 2016, and is long overdue a fix. The motivation
for fixing it now was that it was causing the PackageManager looseMethods
cache to become out of date for the loose class-side methods of classes
mutated by the ClassBuilder, e.g. to add or remove instance variables. At
the top level the cache is an IdentityDictionary indexed by class object,
so for the loose class methods the key is the Metaclass instance. If the
identity of the meta instance is not maintained when modifying a class,
then the fact that some class side methods are extension methods owned
by packages other than the class' owning package will be lost, at least
until the cache is rebuilt.